Moves tests to github actions
