+++ .github/workflows/tests.yml
@@ -1,0 +1,41 @@
+name: tests
+
+on: ['push', 'pull_request']
+
+jobs:
+  linux_tests:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: true
+      matrix:
+        php: [7.1, 7.2, 7.3, 7.4, 8.0]
+        stability: [prefer-stable]
+
+    name: PHP ${{ matrix.php }} - ${{ matrix.stability }}
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Setup PHP
+        uses: shivammathur/setup-php@v2
+        with:
+          php-version: ${{ matrix.php }}
+          tools: composer:v2
+          coverage: none
+
+      - name: Setup Problem Matches
+        run: |
+          echo "::add-matcher::${{ runner.tool_cache }}/php.json"
+          echo "::add-matcher::${{ runner.tool_cache }}/phpunit.json"
+
+      - name: Install PHP 7 dependencies
+        run: composer update --${{ matrix.stability }} --no-interaction --no-progress
+        if: "matrix.php < 8"
+
+      - name: Install PHP 8 dependencies
+        run: composer update --${{ matrix.stability }} --ignore-platform-req=php --no-interaction --no-progress
+        if: "matrix.php >= 8"
+
+      - name: Execute unit tests
+        run: vendor/bin/phpunit --testsuite=Unit
